pants: clean up uses_services pants plugin
